+++ CHANGES.rst
@@ -7,6 +7,14 @@ Changelog of lizard-task
 
 - Import simplejson from django.utils.
 
+- Removed unused fields from TaskExecution model
+
+- Added 'task_uuid' field to TaskExecution model
+
+- Added 'null=True' constraint to 'task' field in TaskExecution model
+
+- Added 'task_uuid' parameter to function get_handler in handler.py
+
 
 0.1 (2012-03-14)
 ----------------